Make graph_craft/core_types dependency optional
