+++ Cargo.toml
@@ -169,6 +169,7 @@ color = "0.3"
 # Linebender ecosystem (END)
 rand = { version = "0.9", default-features = false, features = ["std_rng"] }
 rand_chacha = "0.9"
+rmp-serde = "1.3"
 glam = { version = "0.32.1", default-features = false, features = [
 	"nostd-libm",
 	"scalar-math",
