+++ document/graph-storage/Cargo.toml
@@ -6,16 +6,21 @@ version.workspace = true
 license.workspace = true
 authors.workspace = true
 
+[features]
+conversion = ["dep:graph-craft"]
+default = ["conversion"]
+
 [dependencies]
-serde = { workspace = true }
-serde_json = { workspace = true }
-graph-craft = { workspace = true }
+graph-craft = { workspace = true, optional = true }
 graphene-resource = { workspace = true }
 core-types = { workspace = true }
-blake3 = "1.5"
-rmp-serde = "1.3"
-rustc-hash = "2.0"
-thiserror = "2.0"
+
+thiserror = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+blake3 = { workspace = true }
+rustc-hash = { workspace = true }
+rmp-serde = { workspace = true }
 
 [dev-dependencies]
 graph-craft = { workspace = true, features = ["loading"] }
