+++ document/graph-storage/src/lib.rs
@@ -4,18 +4,22 @@ pub mod attributes;
 pub mod crdt;
 pub mod delta;
 pub mod document;
-pub mod from_runtime;
 pub mod ids;
-pub mod metadata_source;
 pub mod model;
 pub mod registry;
 pub mod resources;
 pub mod session;
+
+#[cfg(any(feature = "conversion", test))]
+pub mod from_runtime;
+#[cfg(any(feature = "conversion", test))]
+pub mod metadata_source;
+#[cfg(any(feature = "conversion", test))]
 pub mod to_runtime;
 
 pub use attributes::*;
 pub use crdt::*;
-pub(crate) use document::*;
+pub use document::*;
 pub use from_runtime::{RuntimeConversion, decode_declaration, encode_declaration};
 pub use ids::*;
 pub use metadata_source::{InputMetadataEntry, NetworkMetadataEntry, NoMetadata, NodeMetadataEntry, NodeMetadataSource};
